route for doctor list help full for admin& change api route add extra folders and some minor change in frontend
